split flintrock into multiple modules
